epaysave_stageone/requirement.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.0.1
django-fernet-fields==0.6
django-filter==2.2.0
django-filters==0.2.1
django-rest-auth==0.9.5
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
drf-writable-nested==0.5.3
idna==2.8
Pillow==6.2.1
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7